Refactor: VC -> View 출력 흐름 리팩토링
